Added support for multi-board testing in CAN tests
* When running CAN tests, the second board (slave) runs supporting code
to verify that CAN messages are received correctly (see the rx_callback
fn).

* Clean up and refinements; style fixes.
