--- examples/assert_can/makefile
@@ -1,2 +1,0 @@
-PROG = assert_can
-include ../makefile
